wave4(W4.0): rename apps/sessions -> apps/session_sharing
Frees the `sessions` name for the live-session (multiplayer chat) harness.
The transcript-sharing app keeps its Django app_label `shared_sessions` (it was
already set explicitly), so this is a PURE Python-package rename: zero DB change
(tables stay `shared_*`, migrations + content-types untouched), no relabel
migration, no prod data move. Routes (/api/sessions, /api/share) and the
frontend are unchanged. Boundary classifier + ARCHITECTURE.md updated.

508 backend tests pass; boundary invariant clean.
