+++ tests/test_architecture_boundary.py
@@ -24,7 +24,7 @@ ROOT = pathlib.Path(__file__).resolve().parent.parent
 APPS = ROOT / "apps"
 
 # ── The tiers (canonical copy lives in ARCHITECTURE.md; keep them in sync) ──────
-FRAMEWORK = {"agents", "agent_runs", "workspaces", "api", "common", "timeline", "tokens", "sessions", "issues", "mcp", "system"}
+FRAMEWORK = {"agents", "agent_runs", "workspaces", "api", "common", "timeline", "tokens", "session_sharing", "issues", "mcp", "system"}
 PRODUCT = {"projects", "collections", "skills", "evals",
            "walkthroughs", "reviews", "shareouts", "runs"}
